feat: Add English skill name translations, courtesy of @Harkains
Skills are now translated from their IDs into English names. This work
builds off of the large amount of work from Harkains' skill translation
file.

https://github.com/Harkains/gbfr-parser/blob/2d070c2ff9a04f37604058bb87c18973856d687c/views/src/en.json

This uses a translation fallback method, in case new IDs are found:

1. skills.{childCharacterType}.{skillID}

This handles cases for Id's dragonform, in which case Dragonform skills
are attributed to the child character ID of "Pl2000", instead of Id's
main ID of "Pl1900".

2. skills.{characterType}.{skillID}

Normal path for how skills are translated from character to character.

3. skills.default.${skillID}

Default skill names for specific skills (if any).

4. skills.default.unknown-skill

If all else fails, then we display unknown skill with the ID so that it
can be translated later on.